requirements.txt
dnspython
cryptography
